requirements.txt
Flask==3.0.3
gunicorn==22.0.0
requests==2.32.3
msal==1.31.0
pandas==2.2.2
openpyxl==3.1.5
google-cloud-storage==2.18.2
